Sample data for demographic and socioeconomic information.

A dataset containing demographic information about individuals, including their unique identifier, age, height, weight, marital status, education level, employment status, and socioeconomic status.

Usage

data("data_rename")

Format

A tibble with 15 rows and 9 variables:

uniqueidentifier

character: Unique identifier for each individual.

where_from

character: The region where the individual is from.

age

numeric: Age of the individual in years.

height

numeric: Height of the individual in meters.

weight

numeric: Weight of the individual in kilograms.

group

character: Blood type group of the individual.

maritalstatus

character: Marital status of the individual.

highesteducationlevel

character: The highest education level attained by the individual.

employed

character: Employment status of the individual.

socioeconomicstatus

character: The socioeconomic status of the individual.

Source

Simulated data for demonstration purposes.

Examples

dta_gtable(data_rename)

  

uniqueidentifier
      	where_from
      	age
      	height
      	weight
      	group
      	maritalstatus
      	highesteducationlevel
      	employed
      	socioeconomicstatus
    

STM/4921
	North East
	50
	1.64
	59
	AB
	Married
	Bachelors
	Yes
	High

STM/4396
	Central
	34
	1.98
	57
	A
	Married
	Masters
	Yes
	Low

STM/7908
	West
	50
	1.95
	84
	A
	Married
	Bachelors
	Yes
	Low

STM/7243
	Central
	39
	1.52
	63
	O
	Single
	Masters
	Yes
	High

STM/4801
	Central
	52
	1.69
	65
	AB
	Single
	Doctorate
	No
	Low

STM/5134
	West
	50
	1.71
	73
	AB
	Married
	Doctorate
	No
	Middle

STM/7138
	West
	35
	1.73
	46
	A
	Single
	Bachelors
	No
	High

STM/6802
	West
	72
	1.98
	70
	B
	Single
	Doctorate
	No
	Low

STM/4420
	West
	42
	1.62
	103
	AB
	Single
	Masters
	No
	Middle

STM/6351
	North East
	40
	1.89
	96
	A
	Other
	Doctorate
	Yes
	High

STM/4933
	West
	38
	1.91
	67
	B
	Married
	Bachelors
	Yes
	High

STM/4303
	West
	37
	1.56
	75
	A
	Single
	Bachelors
	No
	Low

STM/7465
	South
	45
	1.62
	44
	AB
	Married
	Bachelors
	Yes
	Low

STM/4587
	Central
	67
	1.38
	51
	O
	Married
	Bachelors
	No
	Low

STM/5320
	South
	44
	1.37
	63
	AB
	Married
	Bachelors
	No
	Middle