A dictionary with variable mappings for data standardization.

A dataset containing mappings of old values to new values for certain specified variables in a dataset.

Usage

data("dict_misspelled")

Format

A tibble with 29 rows and 3 variables:

variable

character: The variable for which old and new values are mapped.

old

character: The old value to be replaced.

new

character: The new value to replace the old value.

Source

Simulated data for demonstration purposes.

Examples

dta_gtable(dict_misspelled)

  

variable
      	old
      	new
    

region
	southern
	South

	wset
	West

	North-East
	North East

blood_group
	ba
	AB

	ab

	o
	O

	A+
	A

	a

marital_status
	maried
	Married

	single
	Single

	Singel

	Others
	Other

education
	BSC
	Bachelors

	Bsc

	Bachelers

	MSc
	Masters

	PhD
	Doctorate

	Doctoral

ses
	medium
	Middle

	Midle

	Mediam

	low
	Low

	Hihg
	High

.global
	fasle
	No

	N

	n

	no

	yes
	Yes

	y

	tru