Find duplicate rows based on specific columns

dta_duplicates() identifies and returns duplicate rows in a dataframe or tibble, based on the specified columns.

Usage

dta_duplicates(dat, .columns = names(dat))

Arguments

dat

A dataframe or tibble.

.columns

A set of column names to check for duplicates. Defaults to all columns.

Value

A dataframe or tibble containing only the duplicate rows based on the specified column(s).

Examples

# Create a data frame for demonstration

df <- data.frame(
  id = c(14, 20, 12, 32, 14, 23, 15, 12, 30, 14),
  name = c(
   "Mary", "Mark", "Faith", "David", "Mary", "Daniel", "Christine",
   "Johnson", "Elizabeth", "Mary"
  ),
  age = c(21, 18, 25, 17, 21, 24, 21, 19, 20, 21)
)
dta_gtable(df)
#> Error in dta_gtable(df): could not find function "dta_gtable"

# return duplicated rows by all the columns in the data frame

result <- dta_duplicates(df)
dta_gtable(result)
#> Error in dta_gtable(result): could not find function "dta_gtable"

# return duplicated rows by the column `id`

result2 <- dta_duplicates(df, .columns = id)
dta_gtable(result2)
#> Error in dta_gtable(result2): could not find function "dta_gtable"

# A second demo data frame

df2 <- data.frame(
 gender = c(
   "Male", "Female", "Female", "Male", "Male", "Female", "Male", "Female"
 ),
 education = c(
   "Masters", "Bachelor", "Bachelor", "Masters", "Doctorate", "Masters",
   "Bachelors", "Masters"
 ),
 age = c(25, 30, 30, 25, 40, 25, 40, 25)
)
dta_gtable(df2)
#> Error in dta_gtable(df2): could not find function "dta_gtable"

# return duplicated rows based to the columns 'gender' and 'education'

result2 <- dta_duplicates(df2, .columns = c(gender, education))
dta_gtable(result2)
#> Error in dta_gtable(result2): could not find function "dta_gtable"