Create a styled gt table

dta_gtable() creates a gt table from the input data frame and applies a consistent styling with vertical borders, bold column labels, reduced padding, and customized background colors.

Usage

dta_gtable(dat)

Arguments

dat

A data frame that you want to render and style as a gt table.

Value

A styled gt table object.

Examples

data("women")
dta_gtable(women)
#> Error in dta_gtable(women): could not find function "dta_gtable"

data("mtcars")
dta_gtable(head(mtcars, n = 10))
#> Error in dta_gtable(head(mtcars, n = 10)): could not find function "dta_gtable"