Rename columns in a data frame and tibble based on a dictionary

dta_rename() renames columns in a data frame dat using a dictionary dict. The dictionary specifies the old column names and their corresponding new names.

Usage

dta_rename(dat, dict = NULL, .oldnames, .newnames)

Arguments

dat

A data frame whose columns are to be renamed.

dict

A data frame serving as the dictionary, containing the old and new variable names.

.oldnames

A column in dict specifying the old variable names.

.newnames

A column in dict specifying the new variable names.

Value

A tibble with renamed columns.

Details

The function checks the following:

Both dat and dict are data frames or tibbles.

Columns specified in .oldnames exist in dict.

The new variable names .newnames are unique.

There is no overlap between .oldnames and .newnames to avoid conflicts.

If any issues are detected, an error or warning is raised to ensure proper processing.

Examples

data("data_bmi")
dta_gtable(data_bmi)
#> Error in dta_gtable(data_bmi): could not find function "dta_gtable"

# Rename the columns `age` and `weight` to `age_in_years`
# and `weight_in_kgs`

result <- dta_rename(
  dat = data_bmi,
  .oldnames = c("age", "weight"),
  .newnames = c("age_in_years", "weight_in_kgs")
)

dta_gtable(result)
#> Error in dta_gtable(result): could not find function "dta_gtable"

# Rename all columns to `unique_id`, `age_in_years`,
# `height_in_meters` and `weight_in_kgs`. In this case,

new_names <- c(
  "unique_id",
  "age_in_years",
  "height_in_meters",
  "weight_in_kgs"
)
result2 <- dta_rename(
  dat = data_bmi,
  .oldnames = NA,
  .newnames = new_names
)

dta_gtable(result2)
#> Error in dta_gtable(result2): could not find function "dta_gtable"

# Rename from a dictionary - Begin by loading the dataset
# with columns to be renamed.

data("data_rename")
dta_gtable(head(data_rename))
#> Error in dta_gtable(head(data_rename)): could not find function "dta_gtable"

# The dictionary with the columns `oldnames` and
# `newnames` representing the old and new variable names
# respectively

data("dict_rename")
dta_gtable(dict_rename)
#> Error in dta_gtable(dict_rename): could not find function "dta_gtable"

# Perform the renaming

result3 <- dta_rename(
  dat = data_rename,
  dict = dict_rename,
  .oldnames = old,
  .newnames = new
)

dta_gtable(result3)
#> Error in dta_gtable(result3): could not find function "dta_gtable"