Sample data for demographic, technology, and employment information.

A dataset containing demographic information about individuals, including their unique identifier, age, height, weight, marital status, education level, employment status, socioeconomic status, language, technology ownership, and skills in various tools.

Usage

data("data_sample")

Format

A tibble with 10 rows and 19 variables:

id

character: Unique identifier for each individual.

region

character: The region where the individual is from.

age

numeric: Age of the individual in years.

age_group

character: Age group of the individual.

height

numeric: Height of the individual in meters.

weight

numeric: Weight of the individual in kilograms.

blood_group

character: Blood group of the individual.

marital_status

character: Marital status of the individual.

education

character: The highest level of education attained.

employed

character: Employment status of the individual.

ses

character: Socioeconomic status of the individual.

language

character: The primary language spoken by the individual.

phone

character: The type of phone the individual uses.

transport

character: The primary mode of transport used by the individual.

gadgets_owned

character: A comma-separated list of gadgets owned by the individual.

r

character: Whether the individual has skills in R.

python

character: Whether the individual has skills in Python.

sas

character: Whether the individual has skills in SAS.

stata

character: Whether the individual has skills in Stata.

spss

character: Whether the individual has skills in SPSS.

excel

character: Whether the individual has skills in Excel.

Source

Simulated data for demonstration purposes.

Examples

dta_gtable(head(data_sample))

  

id
      	region
      	age
      	age_group
      	height
      	weight
      	blood_group
      	marital_status
      	education
      	employed
      	ses
      	language
      	phone
      	transport
      	gadgets_owned
      	r
      	python
      	sas
      	stata
      	spss
      	excel
    

STM/7539
	Central
	56
	50-59
	1.70
	73
	AB
	Married
	Bachelors
	Yes
	Middle
	Mandarin
	OnePlus
	Bicycle
	Smart TV, Tablet, Desktop Computer, Digital Camera, Smartphone
	No
	No
	No
	No
	No
	Yes

STM/7993
	Central
	46
	40-49
	1.57
	53
	B
	Married
	Bachelors
	No
	Middle
	French
	OnePlus
	Train
	Smartphone, Smart TV
	No
	Yes
	No
	No
	No
	No

STM/7387
	South
	45
	40-49
	1.47
	85
	AB
	Married
	Bachelors
	No
	High
	Arabic
	Samsung
	Car
	Tablet, Smart TV, Smartwatch, Desktop Computer, Laptop
	No
	Yes
	No
	Yes
	Yes
	No

STM/5598
	West
	37
	30-39
	1.67
	77
	AB
	Single
	Bachelors
	Yes
	Middle
	English
	OnePlus
	Bus
	Desktop Computer, Digital Camera, Smartphone, Smart TV, Tablet
	No
	Yes
	No
	Yes
	No
	No

STM/5901
	North East
	45
	40-49
	1.69
	53
	A
	Single
	Bachelors
	Yes
	Low
	Arabic
	OnePlus
	Bus
	Smartwatch, Desktop Computer
	No
	No
	No
	Yes
	Yes
	No

STM/7529
	North East
	51
	50-59
	1.90
	75
	A
	Married
	Bachelors
	No
	Middle
	Mandarin
	Samsung
	Bus
	Desktop Computer, Smartwatch, Tablet
	Yes
	Yes
	No
	No
	No
	No