Create a styled gt table

dta_gtable() creates a gt table from the input data frame and applies a consistent styling with vertical borders, bold column labels, reduced padding, and customized background colors.

Usage

dta_gtable(dat)

Arguments

dat

A data frame that you want to render and style as a gt table.

Value

A styled gt table object.

Examples

data("women")
dta_gtable(women)

  

height
      	weight
    

58
	115

59
	117

60
	120

61
	123

62
	126

63
	129

64
	132

65
	135

66
	139

67
	142

68
	146

69
	150

70
	154

71
	159

72
	164

data("mtcars")
dta_gtable(head(mtcars, n = 10))

  

mpg
      	cyl
      	disp
      	hp
      	drat
      	wt
      	qsec
      	vs
      	am
      	gear
      	carb
    

21.0
	6
	160.0
	110
	3.90
	2.620
	16.46
	0
	1
	4
	4

21.0
	6
	160.0
	110
	3.90
	2.875
	17.02
	0
	1
	4
	4

22.8
	4
	108.0
	93
	3.85
	2.320
	18.61
	1
	1
	4
	1

21.4
	6
	258.0
	110
	3.08
	3.215
	19.44
	1
	0
	3
	1

18.7
	8
	360.0
	175
	3.15
	3.440
	17.02
	0
	0
	3
	2

18.1
	6
	225.0
	105
	2.76
	3.460
	20.22
	1
	0
	3
	1

14.3
	8
	360.0
	245
	3.21
	3.570
	15.84
	0
	0
	3
	4

24.4
	4
	146.7
	62
	3.69
	3.190
	20.00
	1
	0
	4
	2

22.8
	4
	140.8
	95
	3.92
	3.150
	22.90
	1
	0
	4
	2

19.2
	6
	167.6
	123
	3.92
	3.440
	18.30
	1
	0
	4
	4